**2018年4月25日 **
1、权限分类注释，标注可删除权限。
2、增加双开模式的app跟随外部升级的开关。
3、提高app的jni兼容性。
4、提高对app集成其他插件框架的兼容性。
5、增加设置接口，根据包名进行设置。
6、增加Uri的适配范围，支持通过Uri分享和查看文件。
7、修复一个在三星8.0的问题。

**2018年4月20日 **
1、提高对系统自带的app组件兼容性，更好兼容chrome webview，google
service。

+++ VirtualApp/lib/src/main/java/com/lody/virtual/server/am/VActivityManagerService.java
@@ -904,15 +904,17 @@ public class VActivityManagerService extends IActivityManager.Stub {
             int N = map.size();
             while (N-- > 0) {
                 SparseArray<ProcessRecord> uids = map.valueAt(N);
-                for (int i = 0; i < uids.size(); i++) {
-                    ProcessRecord r = uids.valueAt(i);
-                    if (userId != VUserHandle.USER_ALL) {
-                        if (r.userId != userId) {
-                            continue;
+                if(uids != null) {
+                    for (int i = 0; i < uids.size(); i++) {
+                        ProcessRecord r = uids.valueAt(i);
+                        if (userId != VUserHandle.USER_ALL) {
+                            if (r.userId != userId) {
+                                continue;
+                            }
                         }
-                    }
-                    if (r.pkgList.contains(pkg)) {
-                        killProcess(r.pid);
+                        if (r.pkgList.contains(pkg)) {
+                            killProcess(r.pid);
+                        }
                     }
                 }
             }
